+++ extensions/iceworks-material-helper/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.0
 
 - feat: auto react code snippet
+- feat: add actions for PAGES && COMPONENTS view
 
 ## 0.6.0